LayoutTests/imported/w3c/web-platform-tests/css/css-backgrounds/animations/background-size-composition-expected.txt


PASS Compositing: property <background-size> underlying [40px 40px] from add [60px 60px, 260px 260px] to add [160px 160px] at (-0.25) should be [75px 75px, 325px 325px]
PASS Compositing: property <background-size> underlying [40px 40px] from add [60px 60px, 260px 260px] to add [160px 160px] at (0) should be [100px 100px, 300px 300px]
PASS Compositing: property <background-size> underlying [40px 40px] from add [60px 60px, 260px 260px] to add [160px 160px] at (0.25) should be [125px 125px, 275px 275px]
PASS Compositing: property <background-size> underlying [40px 40px] from add [60px 60px, 260px 260px] to add [160px 160px] at (0.5) should be [150px 150px, 250px 250px]
PASS Compositing: property <background-size> underlying [40px 40px] from add [60px 60px, 260px 260px] to add [160px 160px] at (0.75) should be [175px 175px, 225px 225px]
PASS Compositing: property <background-size> underlying [40px 40px] from add [60px 60px, 260px 260px] to add [160px 160px] at (1) should be [200px 200px, 200px 200px]
PASS Compositing: property <background-size> underlying [40px 40px] from add [60px 60px, 260px 260px] to add [160px 160px] at (1.25) should be [225px 225px, 175px 175px]
PASS Compositing: property <background-size> underlying [20% 40%] from add [40px 80px, 180% 160%] to add [80% 40%] at (-0.25) should be [calc(50px + 0%) calc(100px + 30%), 225% 230%]
PASS Compositing: property <background-size> underlying [20% 40%] from add [40px 80px, 180% 160%] to add [80% 40%] at (0) should be [calc(40px + 20%) calc(80px + 40%), 200% 200%]
PASS Compositing: property <background-size> underlying [20% 40%] from add [40px 80px, 180% 160%] to add [80% 40%] at (0.25) should be [calc(30px + 40%) calc(60px + 50%), 175% 170%]
PASS Compositing: property <background-size> underlying [20% 40%] from add [40px 80px, 180% 160%] to add [80% 40%] at (0.5) should be [calc(20px + 60%) calc(40px + 60%), 150% 140%]
PASS Compositing: property <background-size> underlying [20% 40%] from add [40px 80px, 180% 160%] to add [80% 40%] at (0.75) should be [calc(10px + 80%) calc(20px + 70%), 125% 110%]
PASS Compositing: property <background-size> underlying [20% 40%] from add [40px 80px, 180% 160%] to add [80% 40%] at (1) should be [100% 80%, 100% 80%]
PASS Compositing: property <background-size> underlying [20% 40%] from add [40px 80px, 180% 160%] to add [80% 40%] at (1.25) should be [calc(-10px + 120%) calc(-20px + 90%), 75% 50%]
FAIL Compositing: property <background-size> underlying [40px 40px] from replace [100px 100px] to add [160px 160px] at (-0.25) should be [75px 75px, 75px 75px] assert_equals: expected "75px 75px , 75px 75px " but got "75px 75px "
FAIL Compositing: property <background-size> underlying [40px 40px] from replace [100px 100px] to add [160px 160px] at (0) should be [100px 100px, 100px 100px] assert_equals: expected "100px 100px , 100px 100px " but got "100px 100px "
FAIL Compositing: property <background-size> underlying [40px 40px] from replace [100px 100px] to add [160px 160px] at (0.25) should be [125px 125px, 125px 125px] assert_equals: expected "125px 125px , 125px 125px " but got "125px 125px "
FAIL Compositing: property <background-size> underlying [40px 40px] from replace [100px 100px] to add [160px 160px] at (0.5) should be [150px 150px, 150px 150px] assert_equals: expected "150px 150px , 150px 150px " but got "150px 150px "
FAIL Compositing: property <background-size> underlying [40px 40px] from replace [100px 100px] to add [160px 160px] at (0.75) should be [175px 175px, 175px 175px] assert_equals: expected "175px 175px , 175px 175px " but got "175px 175px "
FAIL Compositing: property <background-size> underlying [40px 40px] from replace [100px 100px] to add [160px 160px] at (1) should be [200px 200px, 200px 200px] assert_equals: expected "200px 200px , 200px 200px " but got "200px 200px "
FAIL Compositing: property <background-size> underlying [40px 40px] from replace [100px 100px] to add [160px 160px] at (1.25) should be [225px 225px, 225px 225px] assert_equals: expected "225px 225px , 225px 225px " but got "225px 225px "
FAIL Compositing: property <background-size> underlying [40px 40px] from add [60px 60px] to replace [200px 200px] at (-0.25) should be [75px 75px, 75px 75px] assert_equals: expected "75px 75px , 75px 75px " but got "75px 75px "
FAIL Compositing: property <background-size> underlying [40px 40px] from add [60px 60px] to replace [200px 200px] at (0) should be [100px 100px, 100px 100px] assert_equals: expected "100px 100px , 100px 100px " but got "100px 100px "
FAIL Compositing: property <background-size> underlying [40px 40px] from add [60px 60px] to replace [200px 200px] at (0.25) should be [125px 125px, 125px 125px] assert_equals: expected "125px 125px , 125px 125px " but got "125px 125px "
FAIL Compositing: property <background-size> underlying [40px 40px] from add [60px 60px] to replace [200px 200px] at (0.5) should be [150px 150px, 150px 150px] assert_equals: expected "150px 150px , 150px 150px " but got "150px 150px "
FAIL Compositing: property <background-size> underlying [40px 40px] from add [60px 60px] to replace [200px 200px] at (0.75) should be [175px 175px, 175px 175px] assert_equals: expected "175px 175px , 175px 175px " but got "175px 175px "
FAIL Compositing: property <background-size> underlying [40px 40px] from add [60px 60px] to replace [200px 200px] at (1) should be [200px 200px, 200px 200px] assert_equals: expected "200px 200px , 200px 200px " but got "200px 200px "
FAIL Compositing: property <background-size> underlying [40px 40px] from add [60px 60px] to replace [200px 200px] at (1.25) should be [225px 225px, 225px 225px] assert_equals: expected "225px 225px , 225px 225px " but got "225px 225px "
FAIL Compositing: property <background-size> underlying [auto, contain] from add [100px 150px] to replace [200px 250px] at (-0.25) should be [75px 125px, 75px 125px] assert_equals: expected "75px 125px , 75px 125px " but got "75px 125px "
FAIL Compositing: property <background-size> underlying [auto, contain] from add [100px 150px] to replace [200px 250px] at (0) should be [100px 150px, 100px 150px] assert_equals: expected "100px 150px , 100px 150px " but got "100px 150px "
FAIL Compositing: property <background-size> underlying [auto, contain] from add [100px 150px] to replace [200px 250px] at (0.25) should be [125px 175px, 125px 175px] assert_equals: expected "125px 175px , 125px 175px " but got "125px 175px "
FAIL Compositing: property <background-size> underlying [auto, contain] from add [100px 150px] to replace [200px 250px] at (0.5) should be [150px 200px, 150px 200px] assert_equals: expected "150px 200px , 150px 200px " but got "150px 200px "
FAIL Compositing: property <background-size> underlying [auto, contain] from add [100px 150px] to replace [200px 250px] at (0.75) should be [175px 225px, 175px 225px] assert_equals: expected "175px 225px , 175px 225px " but got "175px 225px "
FAIL Compositing: property <background-size> underlying [auto, contain] from add [100px 150px] to replace [200px 250px] at (1) should be [200px 250px, 200px 250px] assert_equals: expected "200px 250px , 200px 250px " but got "200px 250px "
FAIL Compositing: property <background-size> underlying [auto, contain] from add [100px 150px] to replace [200px 250px] at (1.25) should be [225px 275px, 225px 275px] assert_equals: expected "225px 275px , 225px 275px " but got "225px 275px "
PASS Compositing: property <background-size> underlying [auto 100px, contain] from add [[object Object]] to replace [auto 200px, contain] at (-0.25) should be [auto 75px, contain]
PASS Compositing: property <background-size> underlying [auto 100px, contain] from add [[object Object]] to replace [auto 200px, contain] at (0) should be [auto 100px, contain]
PASS Compositing: property <background-size> underlying [auto 100px, contain] from add [[object Object]] to replace [auto 200px, contain] at (0.25) should be [auto 125px, contain]
PASS Compositing: property <background-size> underlying [auto 100px, contain] from add [[object Object]] to replace [auto 200px, contain] at (0.5) should be [auto 150px, contain]
PASS Compositing: property <background-size> underlying [auto 100px, contain] from add [[object Object]] to replace [auto 200px, contain] at (0.75) should be [auto 175px, contain]
PASS Compositing: property <background-size> underlying [auto 100px, contain] from add [[object Object]] to replace [auto 200px, contain] at (1) should be [auto 200px, contain]
PASS Compositing: property <background-size> underlying [auto 100px, contain] from add [[object Object]] to replace [auto 200px, contain] at (1.25) should be [auto 225px, contain]
